[REFACTOR/#73] LazyListState 확장 함수 LazyListExtension 파일로 이전
